+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.1.1 (2018-07-13)
 * Fix stable publishing detection
 
 ### 1.1.0 (2018-07-13)